Add better aliasing for Ex commands in Vim
`cabbrev` replaces the given string anywhere in the command, which means
that e.g. we can't run `ag` with `cexpr` without it being transformed to
`grep`.

Add a plugin that only does the replacement if the alias appears at the
start of the Ex command.

Also `~/.vim/after` directory to Mackup.