Update nixpkgs/HM to latest: fix Neovim spec
Nixpkgs now generates `init.vim` a different way than it does with Vim,
and said different way is quite hard to reproduce (function args) in
order to get the same Nix store path. What we _really_ care about is
that `~/.config/nvim/reference.vim` has the same content as the wrapped
config, so test that instead.