Remove references to <nixpkgs> and <home-manager>
Replace with `import`s of `default.nix`, which still uses the NIX_PATH
angle bracket paths.

Lays the groundwork for pinning those dependencies.

There's still a reference to `<nixpkgs>` in
`pkgs/heroku-plugins/default.nix`, but that file is autogenerated by
node2nix and we ensure that it refers to our 'pinned' nixpkgs when we
import that file.