Add flake.nix and corresponding lockfile
Includes a script to generate the lockfile, since the latest
flake-enabled version of Nix has issues with paths with spaces in them
(as is the case with the iCloud folder):

With space (no arg passed)
  => "error (BadURL) 'file://<path>' is not a valid URL"

Passing `file://<$PWD URL-encoded>`
  => "error (Error) 'file://<path>' is unsupported"

Passing `git+file://<$PWD URL-encoded>`
  => "error (Error) program 'git' failed with exit code 123"
  => error from Git: "fatal: cannot change to '<URL-encoded path>': No such file or directory"

To get around this, we copy the `flake.nix` (and `flake.lock` if
present) to the Nix store, generate a temporary Git repo and run `nix
flake update` there instead. We then copy the updated `flake.lock` back
to the original repo, or bail out if there are no changes.