Build Nix profile from a single derivation, declaratively
We now don't use `home-manager switch` at all - instead we construct the
HM generation via the HM flake, then link its packages into a buildEnv
which we then set as the current ~/.nix-profile, and run the activation
script ourselves. In order to be able to find the activation script, we
construct _another_ derivation that contains the script and the
buildEnv, then use the path of that derivation (which is outputted by
`nix-build`) to run the activation script and set the current profile.

Unfortunately we can't use the HM `submoduleSupport` flags to stop it
installing packages, since this also changes the `home.profileDirectory`
(which is used for the location of `hm-session-vars.sh` in `.zshrc`
among other things). We also can't override the
`home.activation.installPackages` fragment with `mkForce` or
`mkOverride`, since the module setting is configured be unique. To hack
around this we temporarily override `nix-env` with a Bash function, and
restore it such that HM can still use it to set the current
`home-manager` generation.

Necessary changes:

- Username and $HOME have to be passed in as arguments, as HM's flake
  requires them as arguments.
- The 'host' from which to take extra config is also passed in, which
  means we now longer need `modules/host` (and we now get the serial
  number via a much simpler way).
- For specs, we get the enabled roles via the new `script/switch`
  script, to ensure we get all the necessary arguments injected.
- We need to inject the path of the repo, since it gets copied into the
  Nix store (with only the files that Git knows about), in order to
  still be able to use git-ignored paths for symlinking.