Include JSON file of roles config in HM generation
Means that we can reference it in tests, rather than having to shell-out
to nix to eval the HM config.

Since we now only get the status of roles as-of the last HM generation,
we add an env-var override so that we can treat roles as enabled for the
purposes of writing initially-failing tests.