Remove usage of Nix channels
We still set NIX_PATH to the corresponding sources from flake.nix for
ad-hoc scripting and `nix-shell` shebangs (esp. those in this repo) that
rely on it, but otherwise ensure that we don't use NIX_PATH for building
the user profile, and that we don't have any channels installed that
could sneak in somewhere.