Convert `.macos` script to HM settings
Now that HM has a `defaults` module - which supports arbitrary
domains/keys without having to make a `option` for them - use that
in place of the script.

Not all settings have been ported - only the easier stuff - and guard
settings that need `sudo` so that we only need to enter the password
when a setting changes.