Stop symlinking plists in favour of `defaults`
Introduces problems with clashing versions, particularly with iTerm and
its own 'sync settings' functionality (ended up with multiple files in
different locations and it was unclear which was being used). Preference
plists also tend to include version- and machine-specific info - such as
the date of the last update check - that don't travel well between
machines.

Since all we care about are generally a handful of settings, we set
those directly rather than having to sync the entire file:

- iTerm: added changes seen from diff-ing `defaults read` after deleting
  the whole domain, and then setting preferences again. Changes seen
  that represent default settings are ignored.
- Dash: has its own syncing mechanism which we merely configure.
- Terminal, Symbolic Hotkeys: these are already set in other files.
- 1password: don't change any settings from the defaults, most relevant
  settings seem to be bound up with the account/vault anyway.
- KeepingYouAwake, eqMac: don't change any settings from the defaults.