Change method of CPU type detection
When running under Rosetta 2 (which is the case for processes spawned by
Nix or a Nix-installed shell), `uname -m` returns "x86_64", rather than
the native "arm64" architecture.

There's a couple of ways to get round this:
- `sysctl -in hw.optional.arm64` gives "1" on ARM-native machines,
  regardless of CPU emulation.
- `sysctl -in sysctl.proc_translated` gives "1" when running under
  Rosetta 2 (so would be need to be combined with another arch check).
- `uname -v` contains the CPU type as part of the version string.

Went with the last option, since it's available in Ruby without
shelling-out, and is not too big a change from the existing method.