Always run `mas` under Rosetta 2
Since `mas` gets spawned by `brew`, which is now running as ARM64, macOS
tries to run `mas` as ARM64 as well, which fails since the changes
nixpkgs makes to the executable invalidates the executable signature and
thus Big Sur refuses to run it (reporting "Bad CPU type in executable"
even though it is a universal binary).

We fix this by wrapping the executable in a shell script that runs it
under the correct arch.

This seems to be fixed in nixpkgs for v1.8.2 (the changes to `rpath`
have been removed), so we can remove the wrapper when we upgrade
nixpkgs.