Fix specs (and Nix): disable nixpkgs overlay symlink
Currently a bug in Nix >= 2.16 that means that having an overlay file
breaks any nix-* tool that references `<nixpkgs>`, so disable for now as
we can still use the overlay within our flake.

Also tried splitting the overlays out into separate files (with a
function per-file) but got stuck with a cryptic "a function was found
when a list was expected" error that I couldn't find the source of.

Long-term, would like to get rid of overlays entirely, as it's annoying
to have make sure they get taken into account in all the places we use
Nix, and most of them only concern this flake anyway.