Port telescope.nvim config to Lua
Nicer to work in than Vimscript, particular in this case as
telescope.nvim is a Lua plugin.

Also fix `<leader>up` binding for listing files in $PWD, which was
broken due to a typo ("g" instead of "h").

Requires changes to specs, since Lua functions are rendered by our
current introspection code as a file/line number. However, `desc` are
rendered in the line beneath, so we use that to match against (by
appending the description to the function description).