Disable ability for commands to set iTerm title
AWS session manager does this, and it's annoying because it makes it
look like you're still logged into an instance, even after exiting.
Annoyance outweighs the minor utility of the feature, so disable it.

Because this is a iTerm profile-only setting (unlike things like
colours, which are duplicated in the app defaults and the profile), we
can't use `defaults` without overwriting the whole profile, so we have
to fall back to plist edits instead.