Use `nix profile install` for `script/switch`
Main difference with `nix-env` is that `--set` isn't supported by `nix
profile install`, however-

The state-of-the-art for declarative installation [^1] of a Nix profile
hasn't really changed since we moved to using `script/switch`, but in
practice hasn't been a problem as I use `nix-shell` and `direnv` for
ad-hoc or project-specific packages.

Therefore, we just go back to standard package installation for HM's
packages (i.e. via the activation script) and not install the packages
ourselves. This does mean that `script/switch` becomes less useful vs.
using `home-manager switch`, so we'll probably revert to using that
again.

[^1]: only having the specified packages in the profile and nothing else