Use chezmoi to generate `~/.aws/config`
- Avoids leaking 'secret' info
- Switch to using `credential_process` instead of having to invoke
  `aws-vault exec`
- Set up `mfa_process` for AWS Vault so that we don't need to open
  1Password, copy the OTP, then paste it into the prompt
- Set default region
- Hardcode paths to `op` and `aws-vault` in case an AWS SDK reads the
  config outside of a login shell (i.e. $PATH not set)

Add tests for config file - ideally we'd read these from the `aws`
command itself, but it doesn't have a way to export the entire config
(like `git` does, for example). Since all the AWS CLI does for `aws
configure` is read this file, we instead introduce an INI parser to
ShellLib and read it ourselves.